requirements.txt
aiohttp==3.9.5
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
torch==2.3.0
ta==0.11.0
pyyaml==6.0.1
requests==2.31.0
python-dotenv==1.0.1
